Вынес код тестов в отдельный файл
